[noup] zephyr: Fix fractional part of Tx rate by converting to float
The Tx rate was divided by 1000 using integer division, which
truncated the decimal part (8600 bps → 8 Mbps instead of 8.6 Mbps).
This change ensures the result is stored as a float to retain
precision in Mbps.

+++ wpa_supplicant/wpa_cli_zephyr.h
@@ -18,7 +18,7 @@ struct add_network_resp {
 
 struct signal_poll_resp {
 	int rssi;
-	int current_txrate;
+	float current_txrate;
 };
 
 struct status_resp {
